Moved validation from the main file to helper file
